requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.07.04
cffi==1.16.0
charset-normalizer==3.3.2
chromedrivermanager==0.0.1
colorama==0.4.6
dukpy==0.3.1
et-xmlfile==1.1.0
h11==0.14.0
idna==3.7
iniconfig==2.0.0
openpyxl==3.1.4
outcome==1.3.0.post0
packaging==24.0
pluggy==1.4.0
pycparser==2.21
pypiwin32==223
PySocks==1.7.1
pytest==8.1.1
pywin32==306
requests==2.32.3
selenium==4.18.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.10.0
urllib3==2.2.2
wsproto==1.2.0

PyAutoGUI~=0.9.54